feat(project): 实现存档功能, 添加相关测试, 修复系列问题
- ArchiveManger 修改为 ArchiveIn 类，负责创建存档文件
- 添加 ArchiveContext 类，用于管理存档上下文
- 更新 ArchiveAsset 类，调整字段类型以适应存档需求- 编写单元测试和集成测试，验证存档功能的正确性
- 更新数据库设计文档，明确时间戳和整数类型的使用
- 修复部分数据库记录错误, 添加 tar 文件头对于存档体积的影响

+++ src/main/java/com/cc01cc/project/dto/FileVolumeAsset.java
@@ -22,11 +22,11 @@ import lombok.extern.slf4j.Slf4j;
 @Data
 @Slf4j
 public class FileVolumeAsset {
-    private long id;
-    private long fileId;
-    private long assetId;
-    private int volumeOrder;
+    private Long id;
+    private Long fileId;
+    private Long assetId;
+    private Long volumeOrder;
     private String status;
-    private long createdAt;
-    private long updatedAt;
+    private Long createdAt;
+    private Long updatedAt;
 }
